chore: move the simulation folder out of the unit tests
